Digital Clock
I have added some coding logic to make the format similar to 02:06:45 pm. I have converted time from military to standard. I still need to test every hour of the day to make sure it displays properly. When I set the clock on my laptop to 10 or 11 pm, the display places a 0 before it, which should not happen. I was unable to test this via military time though, which is how the function initiates the time and must be converted after. Further testing is needed.